已修复。所有 DTO 的变更：
变更	说明
Object → BigDecimal	所有 batchQty、totalAmount、planQuantity、quantity、unitPrice、totalValue 字段从 Object 改为 BigDecimal
补齐 Javadoc	所有 DTO 字段加上 /** ... */ 注释，Apifox 可识别字段名和中文描述
DashboardMetricsDto	summary 类型从 Double 改为 BigDecimal
Controller	summary 赋值从 Math.round() 改为 BigDecimal.valueOf()

+++ src/main/java/com/tonghui/erp/Common/Dto/Dashboard/DashboardMetricsDto.java
@@ -1,6 +1,7 @@
 package com.tonghui.erp.Common.Dto.Dashboard;
 
 import lombok.Data;
+import java.math.BigDecimal;
 
 /**
  * 核心指标DTO（含明细）
@@ -20,7 +21,7 @@ public class DashboardMetricsDto {
     /**
      * 预估产值：汇总值（万元）+ 工单明细
      */
-    private MetricsItem<Double, WorkOrderMetricDetailDto> estimatedOutputValue;
+    private MetricsItem<BigDecimal, WorkOrderMetricDetailDto> estimatedOutputValue;
 
     /**
      * 总订单量：汇总值（单）+ 生产计划明细
@@ -35,7 +36,7 @@ public class DashboardMetricsDto {
     /**
      * 总采购额：汇总值（万元）+ 入库单明细
      */
-    private MetricsItem<Double, StockInMetricDetailDto> totalPurchaseAmount;
+    private MetricsItem<BigDecimal, StockInMetricDetailDto> totalPurchaseAmount;
 
     /**
      * 待生产数量：汇总值（批）+ 待生产计划明细
